requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==4.2.7
django-crispy-forms==2.1
django-filter==23.3
django-haystack==3.2.1
djangorestframework==3.14.0
idna==3.6
Markdown==3.5.1
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
Whoosh==2.7.4
